+++ apps/mobile/App.tsx
@@ -4,8 +4,8 @@ import { RelayEnvironmentProvider } from "react-relay";
 import { useState } from "react";
 import { createRelayEnvironment } from "./src/contexts/relay/RelayProvider";
 import { NavigationContainer } from "@react-navigation/native";
-import { MainTabNavigator } from "./src/navigation/MainTabNavigator";
 import { SafeAreaProvider } from "react-native-safe-area-context";
+import { MainTabNavigator } from "~/navigation/MainTabNavigator/MainTabNavigator";
 
 export default function App() {
   const [relayEnvironment] = useState(() => createRelayEnvironment());
