+++ apps/mobile/package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "@react-navigation/material-top-tabs": "^6.6.0",
     "@react-navigation/native": "^6.1.4",
+    "@react-navigation/native-stack": "^6.9.10",
     "@shopify/flash-list": "1.3.1",
     "expo": "~47.0.12",
     "expo-av": "~13.0.2",
@@ -15,6 +16,7 @@
     "react": "18.1.0",
     "react-native": "0.70.5",
     "react-native-fast-image": "^8.6.3",
+    "react-native-gesture-handler": "~2.8.0",
     "react-native-pager-view": "6.0.1",
     "react-native-safe-area-context": "4.4.1",
     "react-native-screens": "~3.18.0",
