+++ apps/mobile/src/navigation/types.ts
@@ -1,0 +1,5 @@
+export type MainStackNavigatorParamList = {
+  Tabs: {};
+  Account: {};
+  Notifications: {};
+};
