+++ apps/mobile/tsconfig.json
@@ -4,11 +4,17 @@
     "baseUrl": ".",
     "jsx": "react-native",
     "paths": {
+      "~/contexts/*": [
+        "src/contexts/*"
+      ],
       "~/frames/*": [
         "../../packages/frames-dls/src/*"
       ],
       "~/generated/*": [
         "./__generated__/*"
+      ],
+      "~/navigation/*": [
+        "src/navigation/*"
       ],
       "~/shared/*": [
         "../../packages/shared/src/*"
